Add parser module and start modeling target AST
The parser module was created to handle the consumption of the tokenized
input produced by the lexer.

Minor code cleanup was performed and additional documentation was added
throughout the project. The aim is to have 100% documentation coverage.
During the prototyping phase, however, this is not a priority.

+++ src/lexer/macros.rs
@@ -79,8 +79,8 @@ macro_rules! kind {
     [int] => {
         $crate::lexer::Kind::Integer
     };
-    [float] => {
-        $crate::lexer::Kind::Float
+    [double] => {
+        $crate::lexer::Kind::Double
     };
     [ident] => {
         $crate::lexer::Kind::Ident
